+++ ThreeDSecure/src/main/java/com/braintreepayments/api/ThreeDSecureRequest.java
@@ -12,6 +12,8 @@ import org.json.JSONObject;
 
 import java.lang.annotation.Retention;
 import java.lang.annotation.RetentionPolicy;
+import java.util.HashMap;
+import java.util.Map;
 import java.util.List;
 
 /**
@@ -56,6 +58,7 @@ public class ThreeDSecureRequest implements Parcelable {
 
     private String nonce;
     private String amount;
+    private Map<String, String> customFields;
     private String mobilePhoneNumber;
     private String email;
     private @ThreeDSecureShippingMethod int shippingMethod;
@@ -80,6 +83,15 @@ public class ThreeDSecureRequest implements Parcelable {
      */
     public void setNonce(@Nullable String nonce) {
         this.nonce = nonce;
+    }
+
+    /**
+     * Set Custom Fields
+     *
+     * @param customFields Object where each key is the name of a custom field which has been configured in the Control Panel. In the Control Panel you can configure 3D Secure Rules which trigger on certain values.
+     */
+    public void setCustomFields(@Nullable Map<String, String> customFields) {
+        this.customFields = customFields;
     }
 
     /**
@@ -298,6 +310,11 @@ public class ThreeDSecureRequest implements Parcelable {
         return email;
     }
 
+    @Nullable
+    public Map<String, String> getCustomFields() {
+        return customFields;
+    }
+
     /**
      * @return The shipping method to use for 3D Secure verification
      */
@@ -429,6 +446,16 @@ public class ThreeDSecureRequest implements Parcelable {
         dest.writeParcelable(v2UiCustomization, flags);
         dest.writeParcelable(v1UiCustomization, flags);
         dest.writeString(accountType);
+        if (customFields != null) {
+            dest.writeInt(customFields.size());
+            for (Map.Entry<String, String> entry: customFields.entrySet()) {
+                dest.writeString(entry.getKey());
+                dest.writeString(entry.getValue());
+            }
+        } else {
+            // set custom fields size == 0 so we know not to create a customFields map when deserializing
+            dest.writeInt(0);
+        }
     }
 
     public ThreeDSecureRequest(Parcel in) {
@@ -448,6 +475,15 @@ public class ThreeDSecureRequest implements Parcelable {
         v2UiCustomization = in.readParcelable(ThreeDSecureV2UiCustomization.class.getClassLoader());
         v1UiCustomization = in.readParcelable(ThreeDSecureV1UiCustomization.class.getClassLoader());
         accountType = in.readString();
+        int customFieldsSize = in.readInt();
+        if (customFieldsSize > 0) {
+            customFields = new HashMap<>();
+            for (int i = 0; i < customFieldsSize; i++) {
+                String key = in.readString();
+                String value = in.readString();
+                customFields.put(key, value);
+            }
+        }
     }
 
     public static final Creator<ThreeDSecureRequest> CREATOR = new Creator<ThreeDSecureRequest>() {
@@ -481,6 +517,11 @@ public class ThreeDSecureRequest implements Parcelable {
 
             if (cardAddChallengeRequested != null) {
                base.put("card_add", cardAddChallengeRequested);
+            }
+
+            if (customFields != null && !customFields.isEmpty()) {
+                JSONObject customFieldsJson = new JSONObject(customFields);
+                base.put("custom_fields", customFieldsJson);
             }
 
             additionalInfo.putOpt("mobile_phone_number", getMobilePhoneNumber());